% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_with_kriegstein_data}
\alias{annotate_with_kriegstein_data}
\title{Perform annotation with chunked Kriegstein reference data}
\usage{
annotate_with_kriegstein_data(
  sample_names,
  sample_files,
  kriegstein_data_dir,
  kriegstein_chunks_input_dir,
  kriegstein_annotated_output_dir,
  annotations = c("age", "structure", "custom.clusterv2")
)
}
\arguments{
\item{sample_names}{Character vector with sample names.}

\item{sample_files}{Character vector with sample .rds files to be annotated.}

\item{kriegstein_data_dir}{String with Kriegstein folder path, containing (custom.)meta.tsv and exprMatrix.tsv.gz.}

\item{kriegstein_chunks_input_dir}{String with Kriegstein input directory for data chunks.}

\item{kriegstein_annotated_output_dir}{String with Kriegstein output directory for annotated data chunks.}

\item{annotations}{default: c("age", "structure", "custom.clusterv2"), see (custom.)meta.tsv for selectable features.}
}
\description{
Perform annotation with chunked Kriegstein reference data
}
\note{
Run EMC.SKlab.scRNAseq::chunk_kriegstein_data() to have chunked reference data before annotating!
No need to run if chunked data already exists!
}
\examples{
kriegstein_data_dir <-  "path/to/Kriegstein_data/"

annotate_with_kriegstein_data(
  sample_names = c("sample_a", "sample_b"),
  sample_files = c("path/to/sample_a.rds", "path/to/sample_b.rds"),
  kriegstein_data_dir = kriegstein_data_dir,
  kriegstein_chunks_input_dir = paste0(kriegstein_data_dir, "/Kriegstein_chunks/"),
  kriegstein_annotated_output_dir = paste0(kriegstein_data_dir, "Kriegstein_annotated_RData/"),
  annotations = c("age", "structure", "custom.clusterv2")
)

}
