% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beep}
\alias{beep}
\title{Beep notification}
\usage{
beep(n = 5)
}
\arguments{
\item{n}{how many beeps to give in succession}
}
\description{
Used for internal testing, get a beep notification, e.g. after long run.
}
