% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{chunk_kriegstein_data}
\alias{chunk_kriegstein_data}
\title{Run Kriegstein reference data chunking for annotation.}
\usage{
chunk_kriegstein_data(
  n_chunks,
  kriegstein_data_dir,
  kriegstein_chunks_output_dir,
  kriegstein_custom_annotation
)
}
\arguments{
\item{n_chunks}{default: 25, integer, amount of data chunks, less memory needed per chunk}

\item{kriegstein_data_dir}{String with Kriegstein folder path, containing meta.tsv and exprMatrix.tsv.gz.}

\item{kriegstein_chunks_output_dir}{String with Kriegstein output directory for data chunks.}
}
\description{
Need to chunk data before training and annoation as to reduce memory needed
per interation. If no chunking needed set n_chunkcs to 1.
}
\note{
Added Kriegstein mapping based on visual inspection of their UCSC
CellBrowser UMAP, see custom_annotation.txt.

Can't train first and then intersect to common genes space to classify after.

Run EMC.SKlab.scRNAseq::chunk_kriegstein_data() to have chunked reference data before annotating!
No need to run if chunked data already exists!
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
kriegstein_custom_annotation <- system.file("extdata", "kriegstein_custom_annotation.txt", package = 'EMC.SKlab.scRNAseq')
chunk_kriegstein_data(
  n_chunks = 25,
  kriegstein_data_dir = "path/to/Kriegstein_data/",
  kriegstein_chunks_output_dir = paste0(kriegstein_data_dir, "RData/chunks/"),
  kriegstein_custom_annotation = kriegstein_custom_annotation
)
\dontshow{\}) # examplesIf}
}
