% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE2volcanoPlot.R
\name{DE2volcanoPlot}
\alias{DE2volcanoPlot}
\title{Visualize DEG in Volcano plot.}
\usage{
DE2volcanoPlot(DE_csv, output_filename, log2FC_threshold, top_n = 5)
}
\arguments{
\item{DE_csv}{Input DEA as .csv format.}

\item{output_filename}{Plot output filename.}

\item{log2FC_threshold}{log2 fold-change threshold for defining DEG.}

\item{top_n}{default: 5. Amount of top up/down regulated genes to plot.}
}
\description{
Visualize DEG in Volcano plot.
}
\examples{
DE_csv <- file.path("DEA_results", "DEG.csv")
output_filename <- "A+C_DEG_volcano-plot_log2FC-threshold=1.png"
log2FC_threshold <- 1
top_n <- 5
DE2volcanoPlot(DE_csv, output_filename, log2FC_threshold, top_n = top_n)
}
