% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DE_EnhancedVolcano}
\alias{DE_EnhancedVolcano}
\title{Perform DE with Seurat FindMarkers() and plot EnhancedVolcano}
\usage{
DE_EnhancedVolcano(
  seurat_object,
  ref_ident,
  vs_ident,
  DE_output_dir,
  comp_name,
  DE_test
)
}
\arguments{
\item{seurat_object}{integrated Seurat object. Containing ref_ident and vs_ident in Idents(seurat_object)}

\item{ref_ident}{reference sample name, pct.1 in Seurat DE result}

\item{vs_ident}{versus sample name, pct.2 in Seurat DE result}

\item{comp_name}{used to handle filenaming and EnhancedVolcano plot title}

\item{directory}{to save plot in, filename is based on reference and versus sample}
}
\description{
Perform DE with Seurat FindMarkers() and plot EnhancedVolcano
}
