% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DE_MarkerExpression}
\alias{DE_MarkerExpression}
\title{Plot marker features as violins and dots per DE comparison}
\usage{
DE_MarkerExpression(seurat_object, features_of_interest, idents, output_dir)
}
\arguments{
\item{seurat_object}{SO with integrated samples}

\item{features_of_interest}{List with named vectors with features (genes)}

\item{idents}{for which DE idents to plot}

\item{output_dir}{output directory for plots, string}
}
\description{
Plot marker features as violins and dots per DE comparison
}
