% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{Do_DE_EnhancedVolcano}
\alias{Do_DE_EnhancedVolcano}
\title{Perform DE with Seurat FindMarkers() and plot EnhancedVolcano}
\usage{
Do_DE_EnhancedVolcano(seurat_object, ref_sample, vs_sample, DE_output_dir)
}
\arguments{
\item{seurat_object}{integrated Seurat object. Containing ref_sample and vs_sample in Idents(seurat_object)}

\item{ref_sample}{reference sample name, pct.1 in Seurat DE result}

\item{vs_sample}{versus sample name, pct.2 in Seurat DE result}

\item{directory}{to save plot in, filename is based on reference and versus sample}
}
\description{
Perform DE with Seurat FindMarkers() and plot EnhancedVolcano
}
