% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\docType{package}
\name{EMC.SKlab.scRNAseq-package}
\alias{EMC.SKlab.scRNAseq}
\alias{EMC.SKlab.scRNAseq-package}
\title{import org.Hs.eg.db org.Hs.eg.db # Organism.HomoSapiens.EntrezGene.DataBase}
\description{
Pipeline for automating scRNAseq analysis in the lab.
}
