% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{fgsea_examine_unmapped_genes}
\alias{fgsea_examine_unmapped_genes}
\title{Examine which and how many genes remain unmapped}
\usage{
fgsea_examine_unmapped_genes(gene_names, dea_ensembl_ids, bitr_ensembl_ids)
}
\arguments{
\item{gene_names}{character vector with gene names}

\item{dea_ensembl_ids}{character vector with gene names from the DEA}

\item{bitr_ensembl_ids}{character vector with gene names from bitr}
}
\description{
What using bitr package for mapping genes from e.g. Ensembl to ENTREZ ids
not all genes might be mapped. This function calculates the percentage of
unmapped genes and plots and saves a distribution of them.
}
