% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_color_palette}
\alias{generate_color_palette}
\title{My custom color palettes.}
\usage{
generate_color_palette(type = 'mixed', n = NULL)
}
\arguments{
\item{type}{character. Can be 'mixed' for mixed colors based on R
qualitative palettes, 'colorblind' for Okabe-Ito palette. For gradients,
see ?hcl.colors and notes. Default: mixed}

\item{n}{integer. How many colors to return.}
}
\value{
Color palette as character vector.
}
\description{
This function returns a color palette, based on featured
qualitative palletes in base R, many are colorblind friendly.
}
\note{
Can also look at creating gradients customy with
colorRampPalette(RColorBrewer::brewer.pal(n = 7, name = "PRGn"))(100)
}
\references{
https://i.stack.imgur.com/H8syw.png. test
https://stackoverflow.com/questions/57153428/r-plot-color-combinations-that-are-colorblind-accessible
display.brewer.all(n=50, exact.n=FALSE)
}
