% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{get_msigdb_term2gene}
\alias{get_msigdb_term2gene}
\title{Get TERM2GENE dataframe for custom MSigDB collection/subcollection terms.}
\usage{
get_msigdb_term2gene(
  organism = "hs",
  collection = "h",
  subcollection = c(),
  id = "EZID",
  version = "latest"
)
}
\arguments{
\item{organism}{Either Homo Sapiems 'hs' or Mus Musculus 'mm'}

\item{collection}{The collection(s) must be one or more from
msigdb::listCollections()}

\item{subcollection}{The subcollection(s) must be one or more from
msigdb::listSubCollections()}

\item{id}{Either gene names 'SYM' or Entrez IDs 'EZID'}

\item{version}{'latest' or one of msigdb::getMsigdbVersions()}
}
\value{
Dataframe with 'term' and 'gene' columns.
}
\description{
The molecular signatures database (MSigDB) contains curated collections and
subcollection of gene sets. These can be used for comparison in GSEA/ORA
analyses.
}
