% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getMeta}
\alias{getMeta}
\title{Get Kriegstein meta with custom annotation mapping}
\usage{
getMeta(kriegstein_data_dir)
}
\arguments{
\item{kriegstein_data_dir}{folder path to Kriegstein data}
}
\value{
Kriegstein meta features with custom clusterv2 mapping.
}
\description{
Get Kriegstein meta with custom annotation mapping
}
