% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{group_tree.adjusted}
\alias{group_tree.adjusted}
\title{Overwrite namespace for enrichplot::group_tree function}
\usage{
group_tree.adjusted(
  hc,
  clus,
  d,
  offset_tiplab,
  nWords,
  label_format_cladelab,
  label_format_tiplab,
  offset,
  fontsize,
  group_color,
  extend,
  hilight,
  cex_category,
  ID_Cluster_mat = NULL,
  geneClusterPanel = NULL,
  align,
  add_tippoint = TRUE,
  align_tiplab = TRUE,
  color = "p.adjust"
)
}
\description{
see enrichplot::group_tree for documentation
}
\details{
Added custom functionality to enrichplot::group_tree, I know playing with
namespace is not proper, however please fill me in on how to do this
without changing namespace!
}
