% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{initialize_DEA_adjusted_functions}
\alias{initialize_DEA_adjusted_functions}
\title{Overwrite namespace for Seurat functions}
\usage{
initialize_DEA_adjusted_functions()
}
\description{
Changed following functions:
see Seurat::FindMarkers.default for documentation
see Seurat:::WilcoxDETest for documentation
see Seurat::Foldchange.default for documentation
}
\details{
Added custom functionality to some Seurat functions, I know playing with
namespace is not proper, however please fill me in on how to do this
without changing namespace!
No functionality is removed!
}
