% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{initialize_FGSEA_group_tree_adjusted}
\alias{initialize_FGSEA_group_tree_adjusted}
\title{Overwrite namespace for enrichplot:::group_tree}
\usage{
initialize_FGSEA_group_tree_adjusted()
}
\description{
see enrichplot:::group_tree for documentation
}
\details{
Added custom functionality to enrichplot::group_tree, I know playing with
namespace is not proper, however please fill me in on how to do this
without changing namespace!
}
