% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{plot_DE_stats}
\alias{plot_DE_stats}
\title{Plot DE stats: histograms and scatterplots of average log2 fold-change and pct.1/pct.2}
\usage{
plot_DE_stats(DE_res, filepath, filename = "DE_stats")
}
\arguments{
\item{DE_res}{DE result from Seurat::FIndMarkers()}

\item{filepath}{directory path to save file}

\item{filename}{filename for .pdf with statplots}
}
\description{
Plot DE stats: histograms and scatterplots of average log2 fold-change and pct.1/pct.2
}
