% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{plot_fgsea_result}
\alias{plot_fgsea_result}
\title{Plot GSEA and ORA results}
\usage{
plot_fgsea_result(
  res,
  res_name,
  plot_n_category,
  dea_ids_lfc,
  output_dir,
  gsea_plot_folder,
  ora_plot_folder
)
}
\arguments{
\item{res}{GSEA or ORA result}

\item{res_name}{name given to GSEA or ORA result}

\item{plot_n_category}{how many categories to plot}

\item{dea_ids_lfc}{named character vector with DEA IDs and logfoldchanges}

\item{output_dir}{folder path for output}

\item{gsea_plot_folder}{output folder basename for GSEA results}

\item{ora_plot_folder}{output folder basename for ORA results}
}
\description{
Create and save all different plot for FGSEA results. Heatplot, goplot,
ridgeplot, gseaplot, dotplot, cnetplot, upsetplot, emapplot and treeplot.
}
