% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run GSEA and ORA.}
\usage{
run_fgsea(
  output_dir,
  dea_result_file,
  cellRanger_ensembl_features,
  lfc_threshold = 1,
  seed = 42,
  swap_GSEA_groups = FALSE,
  use_internal_universe = TRUE,
  p_adjust_method = "BH",
  pval_cutoff = 1,
  qval_cutoff = 1,
  min_gene_set_size_gsea = 10,
  max_gene_set_size_gsea = 500,
  min_gene_set_size_ora = 3,
  max_gene_set_size_ora = NA,
  plot_n_category = 30,
  run_msigdb = FALSE,
  gsea_plot_folder = "gene_set_enrichment_analysis",
  ora_plot_folder = "over_representation_analysis"
)
}
\arguments{
\item{output_dir}{output directory for plots, string}

\item{dea_result_file}{Seurat differential expression analysis result object in .csv}

\item{cellRanger_ensembl_features}{CellRanger .csv file with mapping of Ensembl IDs to Ensembl gene symbols}

\item{lfc_threshold}{ORA differentially expressed genes log fold-change threshold, default: 1}

\item{seed}{set pseudorandomness, default: 42}

\item{swap_GSEA_groups}{invert log fold changes to visually swap upregulation/overrepresentation and downregulation/underrepresentation for GSEA, this is independent of positive/negative with ORA, default: FALSE}

\item{use_internal_universe}{TRUE: use DEA measured genes, FALSE: use clusterProfiler::ORA_db() specific genes, default: TRUE}

\item{p_adjust_method}{p-value multiple testing correction method, default: "BH" (Benjamini-Hochberg)}

\item{pval_cutoff}{GSEA/ORA term taken into account by p-val statistic}

\item{min_gene_set_size_gsea}{minimum gene set size for GSEA, default: 10}

\item{max_gene_set_size_gsea}{maximum gene set size for GSEA, default: 500}

\item{min_gene_set_size_ora}{minimum gene set size for ORA, default: 3}

\item{max_gene_set_size_ora}{maximum gene set size for ORA, default: NA}

\item{plot_n_category}{plot top n most significant terms, default: 30}

\item{qval_cutoffGSEA/ORA}{term taken into account by q-val statistic}
}
\description{
Setting a seed can be useful as FGSEA functions have stochasticity.
}
\details{
Run gene set enrichment and over representation analyses with clusterProfiler using fgsea.
}
\note{
Notes during development

clusterProfiler vignette: https://yulab-smu.top/biomedical-knowledge-mining-book

general intro ORA vs GSEA: https://www.pathwaycommons.org/guide/primers/data_analysis/gsea/

use Entrez IDs with clusterProfiler etc to perform GSEA/ORA analyses with FGSEA

Over-representation (enrichment) analysis (ORA): check DEG vs known biological relationship (via gene set)
will find genes where the difference is large and will fail where the difference is small, but evidenced in coordinated way in a set of related genes
limitations: 1) arbitrary user-defined inclusion criteria, 2) equal importance to each gene, 3) assume gene independance

Gene set enrichment analysis (GSEA): unordered collection of related genes (sets)
Pathway: type of gene set if functional relations are ignored
gene set enrichment analysis (GSEA): aggregates the per gene statistics across genes within a gene set
overcoming ORA limitation, by detecting situations where all genes in a predefined set change in a small but coordinated way

steps: local gene-level statistic: adj-p-val / log fold change,
global gene-set statistics: weighted running sum by ranked-list gene prevalence, Enrichment Score (ES) = max score
, Normalized ES (NES): accounting for gene set size significance testing
& multiple testing correction. Type 1 error: family of hypotheses, chance of false positive (significant p-val),
False Discovery Rate (FDR): fraction of rejected null hypothesis (Type 1 error) that are true (false negatives) -
https://www.pathwaycommons.org/guide/primers/statistics/multiple_testing/ -
https://www.nature.com/articles/nbt1209-1135 -
https://jtd.amegroups.com/article/view/13609/11598
}
\examples{
run_fgsea(
  output_dir = "results/fgsea_GOuniverse_treeplot_KEGG_MSigDB_neurons_TPlabels/",
  dea_result_file = astrocyte_deg.csv,
  cellRanger_ensembl_features = "Ensembl genes.tsv",
)

}
