% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{run_fgsea_gsea}
\alias{run_fgsea_gsea}
\title{Run GSEA}
\usage{
run_fgsea_gsea(
  dea_ids_lfc,
  min_gene_set_size_gsea,
  max_gene_set_size_gsea,
  pval_cutoff,
  p_adjust_method,
  msigdb.hs.h,
  msigdb.hs.c2.cp
)
}
\arguments{
\item{dea_ids_lfc}{named character vector with DEA ids and logfoldchanges}

\item{min_gene_set_size_gsea}{minimum geneSetSize}

\item{max_gene_set_size_gsea}{maximum geneSetSize}

\item{pval_cutoff}{p-value threshold}

\item{p_adjust_method}{p-value adjustment method: BH, ...}

\item{msigdb.hs.h}{dataframe with MsigDB Hallmarks Human gene sets}

\item{msigdb.hs.c2.cp}{dataframe with MsigDB C2 Curated Pathways gene sets}
}
\value{
list with all results
}
\description{
Run GSEA for different databases: GO, KEGG, WP, Reactome Pathways,
DO, DGN & MsigDB.
}
