% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGSEA.R
\name{run_fgsea_ora}
\alias{run_fgsea_ora}
\title{Run ORA}
\usage{
run_fgsea_ora(
  deg_names,
  posneg,
  min_gene_set_size_ora,
  max_gene_set_size_ora,
  pval_cutoff,
  p_adjust_method,
  qval_cutoff,
  universe,
  msigdb.hs.h,
  msigdb.hs.c2.cp
)
}
\arguments{
\item{deg_names}{character vector with deg names}

\item{posneg}{up- or down regulated genes}

\item{min_gene_set_size_ora}{minimum geneSetSize}

\item{max_gene_set_size_ora}{maximum geneSetSize}

\item{pval_cutoff}{p-value threshold}

\item{p_adjust_method}{p-value adjustment method: BH, ...}

\item{qval_cutoff}{q-value threshold}

\item{universe}{DEG/GO gene universe character vector}

\item{msigdb.hs.h}{dataframe with MsigDB Hallmarks Human gene sets}

\item{msigdb.hs.c2.cp}{dataframe with MsigDB C2 Curated Pathways gene sets}
}
\value{
ora_results list with all results
}
\description{
Run ORA for different databases: GO, KEGG, WP, Reactome Pathways,
DO, DGN & MsigDB.
}
