% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_integration.R
\name{run_integration}
\alias{run_integration}
\title{Run sample layers integration}
\usage{
run_integration(so, integration_method)
}
\arguments{
\item{so}{seurat object with layers to be integrated}

\item{integration_method}{default: "RPCA", one of c("RPCA", "CCA", "harmony") "harmony", harmony is run from harmony::RunHarmony(), other are Seurat::IntegrateLayers()}
}
\value{
so
}
\description{
Run sample layers integration
}
