% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_integration.R
\name{samples_integration}
\alias{samples_integration}
\title{Perform Seurat sample integration.}
\usage{
samples_integration(
  sample_files,
  sample_names,
  output_dir,
  ref_sample = NULL,
  perform_cluster_level_selection = TRUE,
  perform_cell_level_selection = FALSE,
  selection_panel = c(),
  selection_percent_expressed = 20
)
}
\arguments{
\item{sample_files}{character vector with paths to .rds data files to be integrated}

\item{sample_names}{character vector with sample names of .rds data files}

\item{output_dir}{Package home directory, used to create output directory for results.}

\item{ref_sample}{default: NULL. Otherwise, a sample within sample_names to be used as CCA reference anchor dataset.}

\item{perform_cluster_level_selection}{perform marker selection based on selection panel and selection percent expressed for each cluster, then reintegrate (note: advise to use either cell or cluster based selection, not both)}

\item{perform_cell_level_selection}{perform marker selection based on selection panel and selection percent expressed for each cell, then reintegrate (note: advise to use either cell or cluster based selection, not both)}

\item{selection_panel}{default: c(). Gene/feature marker panel for selection and reintegration}

\item{selection_percent_expressed}{default: 20. Minimal percentage threshold for each marker in panel to be expressed in order to take whole cluster/cell into account}
}
\description{
Seurat sample integration by Canonical Correlation Analysis.
}
\note{
During development notes
Explored different selectionp panels
- Neurons: c("MAP2", "DCX", "NEUROG2") # RBFOX3 <-> DCX
- Astrocytes: c("VIM", "S100B", "SOX9") # SOX9 <-> FABP7

Calling leidenalg via reticulate to run Leiden algorithm instead of Louvain
algorithm with Seurat::FindClusters() used to work, now it has stopped
working. Still trying to find the cause of this, issued on their Github.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
output_dir <- file.path("EMC-SKlab-scRNAseq", "results")

# files and sample names
sample_files <- c(file.path("EMC-SKlab-scRNAseq", "results", "t1.rds"),
               file.path("EMC-SKlab-scRNAseq", "results", "t2.rds"),
               file.path("EMC-SKlab-scRNAseq", "results", "t3.rds"))
sample_names <- c("t1", "t2", "t3")
selection_panel <- c("GENES", "OF", "INTEREST")
samples_integration(sample_files, sample_names, output_dir,
                    selection_panel = selection_panel)
\dontshow{\}) # examplesIf}
}
