% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thisFilePath}
\alias{thisFilePath}
\title{Get file path of file this function was called from}
\usage{
thisFilePath()
}
\description{
Get file path of file this function was called from
}
\note{
https://stackoverflow.com/questions/47044068/get-the-path-of-current-script
}
\examples{
thisFilePath()
}
