% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{differential_expression_analysis}
\alias{differential_expression_analysis}
\title{Run DEA on processed Seurat object}
\usage{
differential_expression_analysis(
  sample_name,
  qs_file,
  output_dir,
  sample_celltype_DEA = NULL,
  features_of_interest = NULL
)
}
\arguments{
\item{sample_name}{sample name, string}

\item{qs_file}{character string file path to .qs file of processed Seurat object}

\item{output_dir}{output directory for plots, string}

\item{sample_celltype_DEA}{list of sample_celltype comparisons, as exampled}

\item{features_of_interest}{marker features to plot as violins and dots per DE comparison}
}
\description{
Changed following functions in mauritsunkel/EMC-SKlab-seurat:
Seurat::FindMarkers.default
Seurat:::WilcoxDETest
Seurat::Foldchange.default

For benchmarking DE methods see: Junttilla et al 2022 (summarizes both Zimmerman et al 2021/2022 and Squair et al 2021!)
If you have replicate samples: pseudobulk (ROTS with sum > mean) > mixed models (best: MAST_RE) > naive models > latent variable models
If no replicate samples: use Seurat default Wilcoxon rank-sum test (with Presto installed for speed) on SCT normalized assay (@data slot)
-  set: recorrect_umi = FALSE, when running on a subset
AUROC suggests that p-values are generally in order for all methods, however large amount of false positives for naive/latent variable models.
Don't use ComBat for batch normalization for DE. If taking into account batch effects for DE use as a latent variable.
Squair et al 2021: "the central principle underlying valid DE analysis is the ability of statistical methods to account for the intrinsic variability of biological replicates".
}
\details{
Perform differential expression analysis on processed and annotated Seurat object.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
differential_expression_analysis(
  sample_name = "T1",
  qs_file = file.path("EMC-SKlab-scRNAseq", "results", "T1.qs"),
  output_dir = file.path("EMC-SKlab-scRNAseq", "results", 'integrated', 'sample_name')
)

sample_celltype_DEA <- list(
c("SampleA_CelltypeA", "SampleB_CelltypeA"),
c("SampleA_CelltypeB", "SampleB_CelltypeB"))
\dontshow{\}) # examplesIf}
}
