% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{differential_expression_analysis}
\alias{differential_expression_analysis}
\title{Run DEA on processed Seurat object}
\usage{
differential_expression_analysis(sample_name, rds_file, output_dir)
}
\arguments{
\item{sample_name}{sample name, string}

\item{rds_file}{character string file path to .rds file of processed Seurat object}

\item{output_dir}{output directory for plots, string}
}
\description{
Changed following functions in mauritsunkel/EMC-SKlab-seurat:
Seurat::FindMarkers.default
Seurat:::WilcoxDETest
Seurat::Foldchange.default
}
\details{
Perform differential expression analysis on processed and annotated Seurat object.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
differential_expression_analysis(
  sample_name = "T1",
  rds_file = file.path("EMC-SKlab-scRNAseq", "results", "T1.rds"),
  output_dir = file.path("EMC-SKlab-scRNAseq", "results", 'integrated', 'sample_name')
)
\dontshow{\}) # examplesIf}
}
