% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_integration.R
\name{integration_analysis}
\alias{integration_analysis}
\title{Analysis of integrated samples}
\usage{
integration_analysis(
  integrated,
  output_dir,
  sample_names,
  sample_name,
  features_of_interest
)
}
\arguments{
\item{integrated}{Integrated Seurat object}

\item{output_dir}{Package home directory, used to create output directory for results.}
}
\description{
Analysis of integrated samples
}
