% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_integration.R
\name{integration_analysis}
\alias{integration_analysis}
\title{Analysis of integrated samples}
\usage{
integration_analysis(integrated)
}
\arguments{
\item{integrated}{Integrated Seurat object}
}
\description{
Analysis of integrated samples
}
