% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE2enhancedVolcanoPlot.R
\name{plotEnhancedVolcano}
\alias{plotEnhancedVolcano}
\title{Create enhanced Volcano plot from Seurat DE}
\usage{
plotEnhancedVolcano(
  seurat_object,
  seurat_DE,
  ref_ident,
  vs_ident,
  filedir,
  comp_name,
  filesuffix = NULL,
  genes_to_label = NULL,
  label_size = 4
)
}
\arguments{
\item{seurat_object}{integrated Seurat object. Containing ref_ident and vs_ident in Idents(seurat_object)}

\item{seurat_DE}{Seurat DE result, from a FindMarkers() function}

\item{ref_ident}{reference sample name, pct.1 in Seurat DE result}

\item{vs_ident}{versus sample name, pct.2 in Seurat DE result}

\item{filedir}{directory to save plot in, filename is based on reference and versus sample, if NULL: return plot}

\item{comp_name}{used to handle filenaming and EnhancedVolcano plot title}

\item{filesuffix}{default NULL, if string: used to add prefix to filename, useful for seperation of comparisons}

\item{genes_to_label}{default: NULL, if gene vector given, only those are labeled}

\item{label_size}{default: 4, change to control label font size}
}
\description{
Create enhanced Volcano plot from Seurat DE
}
