% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE2enhancedVolcanoPlot.R
\name{plotEnhancedVolcano}
\alias{plotEnhancedVolcano}
\title{Create enhanced Volcano plot from Seurat DE}
\usage{
plotEnhancedVolcano(
  seurat_object,
  seurat_DE,
  ref_sample,
  vs_sample,
  filedir,
  genes_to_label = NULL
)
}
\arguments{
\item{seurat_object}{integrated Seurat object. Containing ref_sample and vs_sample in Idents(seurat_object)}

\item{seurat_DE}{Seurat DE result, from a FindMarkers() function}

\item{ref_sample}{reference sample name, pct.1 in Seurat DE result}

\item{vs_sample}{versus sample name, pct.2 in Seurat DE result}

\item{filedir}{directory to save plot in, filename is based on reference and versus sample}

\item{genes_to_label}{default: NULL, if gene vector given, only those are labeled}
}
\description{
Create enhanced Volcano plot from Seurat DE
}
